jQuery and more
Added methods to HittersController that set @hitters and then renders
partial for the table body.
Updated table body with an id so that it could be targeted easily using
jQuery.
Added routes to match the new methods in the controller.
Refactored table header to more easily access elements by id and class.
Wrote a couple widgets to look for clicking links and updates which
link is shown.